@@ -1,10 +1,10 @@
 package com.idealista.application.service.scorer;
 
 import com.idealista.application.model.Ad;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static com.idealista.util.TestUtil.getAdWithDescription;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class DescriptionTextScorerTest {
 
