@@ -3,19 +3,19 @@ package com.idealista.application.service.scorer;
 import com.idealista.application.model.Ad;
 import com.idealista.application.service.scorer.sizedescriptioncalculator.SizeDescriptionCalculatorFactory;
 import com.idealista.application.service.scorer.sizedescriptioncalculator.SizeDescriptionFlatCalculator;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import static com.idealista.application.model.Typology.FLAT;
 import static com.idealista.util.TestUtil.getAdWithTypology;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class SizeDescriptionScorerTest {
 
     @Mock
