@@ -2,17 +2,17 @@ package com.idealista.application.service.scorer.completeadcalculator;
 
 import com.idealista.application.model.Ad;
 import com.idealista.application.model.Typology;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import static com.idealista.util.TestUtil.getEmptyAd;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.verify;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class CompleteAdDefaultCalculatorTest {
 
     @Mock
