@@ -3,11 +3,11 @@ package com.idealista.application.service.usecase;
 import com.idealista.application.model.Ad;
 import com.idealista.application.port.secondary.AdRepository;
 import com.idealista.application.service.scorer.Scorer;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Arrays;
 import java.util.List;
@@ -15,7 +15,7 @@ import java.util.List;
 import static com.idealista.util.TestUtil.getEmptyAd;
 import static org.mockito.Mockito.*;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class CalculateScoreUseCaseTest {
     @Mock
     private AdRepository adRepository;
