@@ -2,22 +2,22 @@ package com.idealista.application.service.usecase;
 
 import com.idealista.application.model.Ad;
 import com.idealista.application.port.secondary.AdRepository;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Arrays;
 import java.util.List;
 
 import static com.idealista.application.model.AdConstant.IRRELEVANT_SCORE;
 import static com.idealista.util.TestUtil.getEmptyAd;
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class FindIrrelevantAdUseCaseTest {
 
     @Mock
