@@ -8,11 +8,11 @@ import com.idealista.infrastructure.api.mapper.PublicAdMapper;
 import com.idealista.infrastructure.api.mapper.QualityAdMapper;
 import com.idealista.infrastructure.api.model.PublicAd;
 import com.idealista.infrastructure.api.model.QualityAd;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
@@ -21,11 +21,11 @@ import java.util.Collections;
 import java.util.List;
 
 import static com.idealista.util.TestUtil.getEmptyAd;
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class AdsControllerTest {
 
     @Mock
