@@ -3,21 +3,22 @@ package com.idealista.infrastructure.persistence.mapper;
 import com.idealista.application.model.Ad;
 import com.idealista.application.model.Typology;
 import com.idealista.infrastructure.persistence.entity.AdEntity;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-import static com.idealista.util.TestUtil.*;
-import static org.junit.Assert.*;
+import static com.idealista.util.TestUtil.getAdEntityWithTypology;
+import static com.idealista.util.TestUtil.getAdWithTypology;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class AdMapperTest {
 
     @Mock
